requirements.txt
pandas
numpy
matplotlib
typing
seaborn 
dill
sklearn
imblearn
xgboost
pymongo
jupyter
streamlit
gunicorn
ipykernel
python-dotenv
-e .